+++ protected/humhub/docs/CHANGELOG.md
@@ -27,6 +27,7 @@ HumHub Change Log
 - Fix: Incorrect First name & Last name message key (githubjeka)
 - Fix: Do not store complete comments with search index (helnokaly)
 - Fix #2319: Run console application before installation
+- Fix: Directory Knob statistics on included modules (e.g. Enterprise Edition)
 
 
 1.2.0-beta.2 (February 24, 2017)